Remove internal package
This commit removes the "internal" package. Golang now raises errors
when a package imports another package called "internal".
