--- internal/testing/pkg.go
@@ -1,1 +1,0 @@
-package testing
