+++ openstack/identity/v3/groups/results.go
@@ -4,7 +4,6 @@ import (
 	"encoding/json"
 
 	"github.com/gophercloud/gophercloud"
-	"github.com/gophercloud/gophercloud/internal"
 	"github.com/gophercloud/gophercloud/pagination"
 )
 
@@ -52,7 +51,7 @@ func (r *Group) UnmarshalJSON(b []byte) error {
 			return err
 		}
 		if resultMap, ok := result.(map[string]interface{}); ok {
-			r.Extra = internal.RemainingKeys(Group{}, resultMap)
+			r.Extra = gophercloud.RemainingKeys(Group{}, resultMap)
 		}
 	}
