+++ openstack/identity/v3/projects/results.go
@@ -4,7 +4,6 @@ import (
 	"encoding/json"
 
 	"github.com/gophercloud/gophercloud"
-	"github.com/gophercloud/gophercloud/internal"
 	"github.com/gophercloud/gophercloud/pagination"
 )
 
@@ -100,7 +99,7 @@ func (r *Project) UnmarshalJSON(b []byte) error {
 			return err
 		}
 		if resultMap, ok := result.(map[string]interface{}); ok {
-			r.Extra = internal.RemainingKeys(Project{}, resultMap)
+			r.Extra = gophercloud.RemainingKeys(Project{}, resultMap)
 		}
 	}
