+++ openstack/identity/v3/regions/results.go
@@ -4,7 +4,6 @@ import (
 	"encoding/json"
 
 	"github.com/gophercloud/gophercloud"
-	"github.com/gophercloud/gophercloud/internal"
 	"github.com/gophercloud/gophercloud/pagination"
 )
 
@@ -49,7 +48,7 @@ func (r *Region) UnmarshalJSON(b []byte) error {
 			return err
 		}
 		if resultMap, ok := result.(map[string]interface{}); ok {
-			r.Extra = internal.RemainingKeys(Region{}, resultMap)
+			r.Extra = gophercloud.RemainingKeys(Region{}, resultMap)
 		}
 	}
