+++ openstack/identity/v3/services/results.go
@@ -4,7 +4,6 @@ import (
 	"encoding/json"
 
 	"github.com/gophercloud/gophercloud"
-	"github.com/gophercloud/gophercloud/internal"
 	"github.com/gophercloud/gophercloud/pagination"
 )
 
@@ -87,7 +86,7 @@ func (r *Service) UnmarshalJSON(b []byte) error {
 			return err
 		}
 		if resultMap, ok := result.(map[string]interface{}); ok {
-			r.Extra = internal.RemainingKeys(Service{}, resultMap)
+			r.Extra = gophercloud.RemainingKeys(Service{}, resultMap)
 		}
 	}
